Type validating smr protected pointer accessors.
This API is intended to provide some measure of safety with SMR
protected pointers.  A struct wrapper provides type checking and
a guarantee that all access is mediated by the API unless abused.  All
modifying functions take an assert as an argument to guarantee that
the required synchronization is present.

Reviewed by:	kib, markj, mjg
Differential Revision:	https://reviews.freebsd.org/D23711


git-svn-id: https://svn.freebsd.org/base/head@358096 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f