MFC r352247-r352248
r352247:
Remove useless extra definition of libedit flags

Note that all the line editor part is done in the libntp

r352248:
Get the readline header from the installed header instead of the from the source
location.

With newer import of libedit, the path to be able to access readline/readline.h
will also include header which name will conflict with some expected by ntp in
another path and end up breaking the build.


git-svn-id: https://svn.freebsd.org/base/stable/12@358110 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f