ncurses: bump shlib number to version 9
ABI has change in between ncurses 5 or 6. While theorically ncurses 6 is buildable with
backward compatibility, I fail at building in a way where the application linked against
the previous version of ncurses are rendering properly.
Let's go on the new ABI which provides all the latest features.

A compat12x package is cooking for backward compatibility


git-svn-id: https://svn.freebsd.org/base/head@358164 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f