MFC r358013: Fix kernel panic while trying to read multicast stream.
When VIMAGE is enabled make sure the "m_pkthdr.rcvif" pointer is set
for all mbufs being input by the IGMP/MLD6 code. Else there will be a
NULL-pointer dereference in the netisr code when trying to set the
VNET based on the incoming mbuf. Add an assert to catch this when
queueing mbufs on a netisr to make debugging of similar cases easier.

Found by:	Vladislav V. Prodan
PR:		244002
Reviewed by:	bz@
Sponsored by:	Mellanox Technologies


git-svn-id: https://svn.freebsd.org/base/stable/12@358273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f