ip6: retire in6_selectroute_fib() as promised 8 years ago
In r231852 I added in6_selectroute_fib() as a compat function with the
fibnum as an extra argument compared to in6_selectroute() to keep the
KPI stable.
Way too late retire this function again and add the fib to in6_selectroute()
which also only has a single consumer now and was an orphan function before.


git-svn-id: https://svn.freebsd.org/base/head@358575 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f