Delete OpenBSD's variant of RCS.
I mistakenly borught this on r282572 but it was never used.
There was some interest for a replacement of the GNU version but
unfortunately the code never passed the tests in the GNU testsuite
and seemed to be poorly maintained. Eventually we moved the GNU
version to the ports tree and ovided carerying RCS in base.


git-svn-id: https://svn.freebsd.org/base/vendor/OpenBSD/dist@358868 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f