MFC r355730:
cxgbe(4): Use the _XT variant of the CPL used to transmit NIC traffic.

CPL_TX_PKT_XT disables the internal parser on the chip and instead
relies on the driver to provide the exact length of the L2 and L3
headers.  This allows hw checksumming and TSO to be used with L2 and
L3 encapsulations that the chip doesn't understand directly.

Note that netmap tx still uses the old CPL as it never uses the hw
to generate the checksum on tx.

Reviewed by:	jhb@
Sponsored by:	Chelsio Communications
Differential Revision:	https://reviews.freebsd.org/D22788


git-svn-id: https://svn.freebsd.org/base/stable/12@358975 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f