Implement a workaround for kms-drm modules
pci_iov_if.h was added to pci.h, but none of the kms-drm branches have
that. Rather than play whack a mole with the branches, move its inclusion to
linux_pci.c which is the only part of the code that needs it now.

Longer term, other solutions will be needed, but this gives us time to get those
deployed on all the supported versions.


git-svn-id: https://svn.freebsd.org/base/head@359165 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f