Use -F instead of -f in tail(1) examples
There is an example in tail(1) manual page explaining how to use tail(1) to
track the contents of /var/log/messages. The example uses the -f flag to
follow the file. The problem with the -f flag is that it cannot handle the
situation where /var/log/messages is rotated. Hence, use -F instead in the
example.

Reviewed by:	asomers
MFC after:	3 days
Differential Revision:	https://reviews.freebsd.org/D24157


git-svn-id: https://svn.freebsd.org/base/head@359229 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f